fix: critical bugs from audit — null return, div-by-zero, UB signed shift - aplayer_camera_manager.hpp:13 — null check had empty body '{};' instead   of 'return {};' causing reads from phys addr 0 when camera invalid (CRASH) - render.hpp:268 — HealthBar div-by-zero guard when v_max==0 (CRASH) - ue_object.hpp:120 — signed int32 key >> 16 is implementation-defined on   negative values; cast to uint32 before shift (MAJOR UB) - aactor.hpp:96 — validate bone_array is a kernel-range address before   reading (MAJOR)

+++ VEX/include/game/sdk/ue_object.hpp
@@ -117,7 +117,7 @@ namespace sky::game::sdk {
             }
 
             // Direct name pool read
-            const std::uint32_t chunkOffset = static_cast<std::uint32_t>(key >> 16);
+            const std::uint32_t chunkOffset = static_cast<std::uint32_t>(static_cast<std::uint32_t>(key) >> 16);
             const std::uint16_t nameOffset = static_cast<std::uint16_t>(key & 0xFFFF);
             const std::uintptr_t chunkAddr = sky::driver::g_driver->get_base_address() + offsets::FNamePool + ((static_cast<unsigned long long>(chunkOffset) + 2ULL) * 8ULL);
             const std::uintptr_t namePoolChunk = sky::driver::g_driver->read<std::uintptr_t>(chunkAddr);
